horizont-portal/mdm/requirement.txt
appnope==0.1.0
Babel==2.5.3
backcall==0.1.0
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
Django==2.0.4
django-babel==0.6.2
django-cors-headers==2.2.0
django-js-asset==1.1.0
django-modelcluster==4.1
django-mptt==0.9.0
django-prices==1.0.0
django-rest-swagger==2.1.2
django-taggit==0.22.2
django-treebeard==4.2.1
django-versatileimagefield==1.9
djangorestframework==3.8.2
djangorestframework-recursive==0.1.2
draftjs-exporter==2.0.0
html5lib==0.999999999
idna==2.6
ipython==6.3.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.12.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.0.0
prices==1.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2018.4
requests==2.18.4
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
text-unidecode==1.1
traitlets==4.3.2
typing==3.6.4
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.22
wagtail==2.0.1
wcwidth==0.1.7
webencodings==0.5.1
Willow==1.1
